refactor(db): Split patch.rs into directory module
Split the 1,659-line patch.rs into a directory module with 6 focused
files following the merge_handler pattern. No behavioral changes.
